src/third_party/blink/web_tests/external/wpt/css/css-scroll-snap/snap-after-relayout/move-current-target-expected.txt

This is a testharness.js-based test.
PASS Moving the current snap target should make the scroller resnap to it.
PASS Changing the layout of other elements should be able to cause resnapping to the target.
FAIL Transforming the current snap target should make the scroller resnap to it. assert_equals: expected 200 but got 100
PASS Applying two property changes that do not change the visual offset of the target should not change the scroll offset.
Harness: the test ran to completion.
